Update senior-engineer-checklist.json
I made some copywriting fixes, and other suggestions. On the "major project" one, I think it's good to emphasize that you should actually try to understand the business's needs and work on something impactful, rather than immediately pouncing on any opportunity to "make your mark." On the "gain influence" one, gaining influence is obviously good for your career but a more positive way to phrase it would be to grow relationships. I don't necessarily fully agree with all the other ones but this is your list and I don't want to bulldoze it. :)